Fix display garbage: restore CurrentScores to unsigned long
The memory optimization that changed CurrentScores from unsigned long (4 bytes)
to unsigned int (2 bytes) broke the display, causing garbage numbers to appear.
The display and score handling expect 32-bit values; 16-bit truncation causes
memory alignment issues and corrupted display output.

Revert only the CurrentScores type change; keep other memory optimizations
for threeBankCompleteCount, fiveBankCompleteCount, and spinnerHitCount.

+++ LectronamoRecharged.ino
@@ -285,7 +285,7 @@ unsigned long BonusCollectionEndTime = 0;
 boolean ExtraBallLaneEnabled = true;
 byte SpecialAwardType = 2;
 
-unsigned int CurrentScores[RPU_NUMBER_OF_PLAYERS_ALLOWED];
+unsigned long CurrentScores[RPU_NUMBER_OF_PLAYERS_ALLOWED];
 unsigned long BallFirstSwitchHitTime = 0;
 unsigned long BallTimeInTrough = 0;
 unsigned long GameModeStartTime = 0;
